+++ R/check_credentials.R
@@ -17,14 +17,30 @@ check_credentials <- function(x) {
   checkmate::reportAssertions(checks)
 
   # PERFORM CHECK
-  if (!is.element(tolower(x), tolower(keyring::key_list()[, 1]))) {
-    res <- paste0("The credentials (username and password) for the service '",
-                  x,
-                  "' have not been saved in the user's profile at this computer. ",
-                  "Use 'NVIdb::set_credentials' to save the credentials in the user's profile")
-  } else {
-    res <- TRUE
+  if (Sys.info()["sysname"] == "Windows") {
+    if (!is.element(tolower(x), tolower(keyring::key_list()[, 1]))) {
+      res <- paste0("The credentials (username and password) for the service '",
+                    x,
+                    "' have not been saved in the user's profile at this computer. ",
+                    "Use 'NVIdb::set_credentials' to save the credentials in the user's profile")
+    } else {
+      res <- TRUE
+    }
   }
+
+  if (Sys.info()["sysname"] == "Linux") {
+    env_variables <- Sys.getenv()
+    if (!any(grepl(paste0(x, ":"), names(env_variables), ignore.case = TRUE) == TRUE)) {
+    # if (!is.element(tolower(x), tolower(keyring::key_list()[, 1]))) {
+      res <- paste0("The credentials (username and password) for the service '",
+                    x,
+                    "' have not been saved for any user at this computer. ",
+                    "Use 'NVIdb::set_credentials' to save the credentials")
+    } else {
+      res <- TRUE
+    }
+  }
+
   return(res)
 }