+++ R/check_odbc_channel.R
@@ -8,9 +8,9 @@
 #'
 #' @templateVar fn odbc_channel
 #' @template x
-#' @param dbservice \[character(1)\]\cr
+#' @param dbservice [\code{character(1)}]\cr
 #'     The database the channel should be connect to.
-#' @param dbinterface \[character(1)\]\cr
+#' @param dbinterface [\code{character(1)}]\cr
 #'     The R-package that is used for interface towards the data
 #'     base. Defaults to \code{NULL}.
 #' @template checker